har fjernet HiddenIf-komponent da den hadde få bruksområder og introduserte unødvendig kompleksitet i kodebasen.
