+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/Skolepenger/InnvilgetSkolepenger/SkoleårsperioderSkolepenger.tsx
@@ -12,8 +12,9 @@ import SkoleårDelårsperiode from './SkoleårDelårsperiode';
 import UtgiftsperiodeSkolepenger from './UtgiftsperiodeSkolepenger';
 import { tomSkoleårsperiodeSkolepenger } from '../typer';
 import navFarger from 'nav-frontend-core';
-import FjernKnappMedTekst from '../../../../../Felles/Knapper/FjernKnappMedTekst';
 import { oppdaterValideringsfeil } from '../utils';
+import { Delete } from '@navikt/ds-icons';
+import { Button } from '@navikt/ds-react';
 
 const Skoleårsperiode = styled.div`
     margin: 1rem;
@@ -118,12 +119,14 @@ const SkoleårsperioderSkolepenger: React.FC<Props> = ({
                                 låsteUtgiftIder={låsteUtgiftIder}
                             />
                             {skalViseFjernKnapp && (
-                                <div>
-                                    <FjernKnappMedTekst
-                                        onClick={() => fjernSkoleårsperiode(index)}
-                                        knappetekst="Fjern skoleår"
-                                    />
-                                </div>
+                                <Button
+                                    onClick={() => fjernSkoleårsperiode(index)}
+                                    type="button"
+                                    variant={'secondary'}
+                                    icon={<Delete title={'Fjern skoleår'} />}
+                                >
+                                    <span>Fjern skoleår</span>
+                                </Button>
                             )}
                         </Skoleårsperiode>
                     </>
