+++ src/frontend/Komponenter/Behandling/Vurdering/Begrunnelse.tsx
@@ -2,10 +2,7 @@ import * as React from 'react';
 import { FC } from 'react';
 import { BegrunnelseRegel, Regel } from './typer';
 import { Vurdering } from '../Inngangsvilkår/vilkår';
-import hiddenIf from '../../../Felles/HiddenIf/hiddenIf';
-import { Textarea as TextareaNav } from '@navikt/ds-react';
-
-const Textarea = hiddenIf(TextareaNav);
+import { Textarea } from '@navikt/ds-react';
 
 interface Props {
     svar: Vurdering;
@@ -17,6 +14,10 @@ const Begrunnelse: FC<Props> = ({ svar, onChange, regel }) => {
     const begrunnelsetype = svar.svar && regel.svarMapping[svar.svar].begrunnelseType;
     const skjulBegrunnelse = (begrunnelsetype ?? BegrunnelseRegel.UTEN) === BegrunnelseRegel.UTEN;
 
+    if (skjulBegrunnelse) {
+        return null;
+    }
+
     return (
         <Textarea
             label={'Begrunnelse '.concat(
@@ -24,7 +25,6 @@ const Begrunnelse: FC<Props> = ({ svar, onChange, regel }) => {
             )}
             value={svar.begrunnelse || ''}
             onChange={(e) => onChange(e.target.value)}
-            hidden={skjulBegrunnelse}
             maxLength={0}
         />
     );
